requirements.txt
plaintext
python-telegram-bot
pillow
pytesseract
gspread
oauth2client
requests